+++ packages/engine/src/index.ts
@@ -1,5 +1,5 @@
 export type * from "./types";
-export * from "./validatePoll";
-export * from "./validateChannel";
+export * from "./types/poll";
+export * from "./types/channel";
 export * from "./questionTypes";
 export * from "./ranking-systems";
