+++ packages/engine/src/types/channel.ts
@@ -1,46 +1,86 @@
-import type { ContentIdentifier, UserId } from "./identifiers";
-import type { PollItem } from "./poll";
+import { pluginField } from "@marcianosrs/form";
+import { FirebaseBaseDTO } from "../types";
+import type { ContentIdentifier } from "./identifiers";
+import { rankingSystemStore } from "../ranking-systems";
+import { FormDataObject, TypedForm } from "@marcianosrs/form-schema";
+import { themeStore } from "@marcianosrs/ui";
 
-export type Channel = {
-    id: ContentIdentifier;
-    name: string;
-    description: string;
-    owner: UserId;
-    theme: ContentIdentifier;
-    scoreSystems: ContentIdentifier[];
-    moderatorIds: string[];
-    createdAt: number;
-    createdBy: string;
-    playlist: PollItem[];
-    frequency: "daily" | "weekly";
-    order: "shuffle" | "asc" | "desc";
-};
+const rankingSystemSchema = [
+	pluginField(
+		"ranking",
+		"Ranking system",
+		rankingSystemStore,
+		"rankingSystemType",
+		"editForm"
+	),
+] as const satisfies TypedForm;
 
-export type Season = {
-    id: ContentIdentifier;
-    name: string;
-    startDate: Date;
-    endDate: Date;
-    theme: ContentIdentifier;
-    scoreSystems: ContentIdentifier[];
-};
+export const channelSchema = [
+	{
+		name: "name",
+		valueType: "string",
+		fieldType: "text",
+		defaultValue: "",
+		displayName: "name",
+		optional: false,
+	},
+	pluginField("theme", "Theme", themeStore, "displayName", "editForm"),
+	{
+		name: "rankingSystems",
+		displayName: "Ranking Systems",
+		fieldType: "objectList",
+		valueType: "objects",
+		objectSchema: rankingSystemSchema,
+		optional: false,
+		minimalAmount: 1,
+	},
+] as const satisfies TypedForm;
 
-export type ProductDefinition = {
-    id: ContentIdentifier;
-    channelId: ContentIdentifier;
-    seasonId?: ContentIdentifier;
-    name: string;
-    description: string;
-    // image?
+export type ChannelDTO = FirebaseBaseDTO &
+	FormDataObject<typeof channelSchema> & { slug: string };
+export type CreateChannelDTO = Omit<ChannelDTO, "id" | "createdAt" | "slug">;
+export type UpdateChannel = Partial<ChannelDTO>;
 
-    theme?: ContentIdentifier;
-    scoreMutators?: ContentIdentifier[];
-    questionModifier?: ContentIdentifier[];
+// export type Channel = {
+//     id: ContentIdentifier;
+//     name: string;
+//     description: string;
+//     owner: UserId;
+//     theme: ContentIdentifier;
+//     scoreSystems: ContentIdentifier[];
+//     moderatorIds: string[];
+//     createdAt: number;
+//     createdBy: string;
+//     playlist: PollItem[];
+//     frequency: "daily" | "weekly";
+//     order: "shuffle" | "asc" | "desc";
+// };
 
-    availableFrom?: Date;
-    availableTill?: Date;
-    availableAmount?: number;
-};
+// export type Season = {
+//     id: ContentIdentifier;
+//     name: string;
+//     startDate: Date;
+//     endDate: Date;
+//     theme: ContentIdentifier;
+//     scoreSystems: ContentIdentifier[];
+// };
+
+// export type ProductDefinition = {
+//     id: ContentIdentifier;
+//     channelId: ContentIdentifier;
+//     seasonId?: ContentIdentifier;
+//     name: string;
+//     description: string;
+//     // image?
+
+//     theme?: ContentIdentifier;
+//     scoreMutators?: ContentIdentifier[];
+//     questionModifier?: ContentIdentifier[];
+
+//     availableFrom?: Date;
+//     availableTill?: Date;
+//     availableAmount?: number;
+// };
 
 /**
  * TODO: Not sure if this is needed, but could be doing things like removing time limits,
@@ -49,9 +89,9 @@ export type ProductDefinition = {
  * This is about ways to answer the question, the score is how to process the result
  */
 export type QuestionModifier<T extends Record<string, unknown>> = {
-    id: ContentIdentifier;
-    name: string;
-    internalMutatorId: string;
+	id: ContentIdentifier;
+	name: string;
+	internalMutatorId: string;
 
-    config: T; // Settings like time-limit
+	config: T; // Settings like time-limit
 };
