+++ packages/frontend/app/routes/admin.channels.new.tsx
@@ -1,54 +1,18 @@
-import { type CreateChannelDTO, rankingSystemStore } from "@marcianosrs/engine";
-import { SchemaForm, pluginField, schemaToZod } from "@marcianosrs/form";
-import type { TypedForm } from "@marcianosrs/form-schema";
+import { type CreateChannelDTO, channelSchema } from "@marcianosrs/engine";
+import { SchemaForm, schemaToZod } from "@marcianosrs/form";
 import type { ActionFunction, LoaderFunction } from "@remix-run/node";
 import { Link } from "@remix-run/react";
 import { makeDomainFunction } from "domain-functions";
 import { formAction } from "../form-action.server";
 import { throwIfNotAuthorized } from "../util/isAuthorized";
 import { createChannel } from "./api.server";
-import { themeStore } from "@marcianosrs/ui";
 
 export const loader: LoaderFunction = async ({ request }) => {
 	await throwIfNotAuthorized(request);
 	return {};
 };
 
-const rankingSystemSchema = [
-	pluginField(
-		"ranking",
-		"Ranking system",
-		rankingSystemStore,
-		"rankingSystemType",
-		"editForm"
-	),
-] as const satisfies TypedForm;
-
-const schema = [
-	{
-		name: "name",
-		valueType: "string",
-		fieldType: "text",
-		defaultValue: "",
-		displayName: "name",
-		optional: false,
-	},
-	// 	{name: "slugDescription",
-	// 	fieldType: "description"
-	// },
-	pluginField("theme", "Theme", themeStore, "displayName", "editForm"),
-	{
-		name: "rankingSystems",
-		displayName: "Ranking Systems",
-		fieldType: "objectList",
-		valueType: "objects",
-		objectSchema: rankingSystemSchema,
-		optional: false,
-		minimalAmount: 1,
-	},
-] as const satisfies TypedForm;
-
-const zodSchema = schemaToZod(schema);
+const zodSchema = schemaToZod(channelSchema);
 
 const mutation = (userId: string) =>
 	makeDomainFunction(zodSchema)(async (values) => {
@@ -75,7 +39,7 @@ export default function NewChannel() {
 		<main>
 			<Link to="/polls">Back to Channel list (homepage?)</Link>
 			<h1>Create new channel</h1>
-			<SchemaForm schema={schema} formId="channel" />
+			<SchemaForm schema={channelSchema} formId="channel" />
 		</main>
 	);
 }
