+++ packages/frontend/app/routes/polls.new.tsx
@@ -1,7 +1,6 @@
 import type { CreatePollDTO } from "@marcianosrs/engine";
-import { questionTypeStore } from "@marcianosrs/engine";
-import { SchemaForm, pluginField, schemaToZod } from "@marcianosrs/form";
-import type { TypedForm } from "@marcianosrs/form-schema";
+import { pollSchema } from "@marcianosrs/engine";
+import { SchemaForm, schemaToZod } from "@marcianosrs/form";
 import type { ActionFunction, LoaderFunction } from "@remix-run/node";
 import { NavLink } from "@remix-run/react";
 import { makeDomainFunction } from "domain-functions";
@@ -14,17 +13,7 @@ export const loader: LoaderFunction = async ({ request }) => {
 	return {};
 };
 
-const schema = [
-	pluginField(
-		"question",
-		"Question",
-		questionTypeStore,
-		"displayName",
-		"editForm"
-	),
-] as const satisfies TypedForm;
-
-const zodSchema = schemaToZod(schema);
+const zodSchema = schemaToZod(pollSchema);
 
 const mutation = (userId: string) =>
 	makeDomainFunction(zodSchema)(async (values) => {
@@ -51,7 +40,7 @@ export default function NewPoll() {
 		<main>
 			<NavLink to="/polls">Back to list of polls</NavLink>
 			<h1>Create new poll</h1>
-			<SchemaForm schema={schema} formId="questionType" />
+			<SchemaForm schema={pollSchema} formId="questionType" />
 		</main>
 	);
 }
